Separating adding books from the main page
